refactor: #2095 change invite new member modal
