--- packages/developer-portal/src/components/ui/developer-invite-member-modal/__tests__/developer-invite-member-footer.test.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react'
-import { shallow } from 'enzyme'
-import { DeveloperInviteMemberModalFooter } from '../developer-invite-member-footer'
-
-describe('DeveloperInviteMemberModalFooter', () => {
-  it('should match snapshot', () => {
-    const wrapper = shallow(<DeveloperInviteMemberModalFooter afterClose={jest.fn()} />)
-    expect(wrapper).toMatchSnapshot()
-  })
-})
