--- packages/developer-portal/src/components/ui/developer-invite-member-modal/__tests__/developer-invite-member-modal-input.test.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react'
-import { shallow } from 'enzyme'
-import { DeveloperInviteMemberModalInput } from '../developer-invite-member-modal-input'
-
-describe('DeveloperInviteMemberModalInput', () => {
-  it('should match snapshot', () => {
-    const wrapper = shallow(<DeveloperInviteMemberModalInput />)
-    expect(wrapper).toMatchSnapshot()
-  })
-})
