+++ packages/developer-portal/src/components/ui/developer-invite-member-modal/__tests__/developer-invite-member-modal.test.tsx
@@ -1,15 +1,22 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
-import { DeveloperInviteMemberModal } from '../developer-invite-member-modal'
+import { InviteMemberModal, InviteMemberModalInput } from '../developer-invite-member-modal'
 
-describe('DeveloperInviteMemberModal', () => {
-  it('should match snapshot with default', () => {
-    const wrapper = shallow(<DeveloperInviteMemberModal afterClose={jest.fn()} />)
-    expect(wrapper).toMatchSnapshot()
+describe('developer-invite-member-modal', () => {
+  describe('InviteMemberModal', () => {
+    it('should match snapshot with default', () => {
+      const wrapper = shallow(<InviteMemberModal afterClose={jest.fn()} />)
+      expect(wrapper).toMatchSnapshot()
+    })
+    it('should match snapshot with visible true', () => {
+      const wrapper = shallow(<InviteMemberModal visible={true} afterClose={jest.fn()} />)
+      expect(wrapper).toMatchSnapshot()
+    })
   })
-
-  it('should match snapshot with visible true', () => {
-    const wrapper = shallow(<DeveloperInviteMemberModal visible={true} afterClose={jest.fn()} />)
-    expect(wrapper).toMatchSnapshot()
+  describe('InviteMemberModalInput', () => {
+    it('should match snapshot', () => {
+      const wrapper = shallow(<InviteMemberModalInput />)
+      expect(wrapper).toMatchSnapshot()
+    })
   })
 })
