--- packages/developer-portal/src/components/ui/developer-invite-member-modal/constants.ts
@@ -1,1 +1,0 @@
-export const MAX_MESSAGE_LENGTH = 150
