--- packages/developer-portal/src/components/ui/developer-invite-member-modal/handlers.ts
@@ -1,4 +1,0 @@
-export const handleSubmit = values => {
-  // TBC
-  console.log('submit', values)
-}
