--- packages/developer-portal/src/styles/blocks/developer-invite-member.scss
@@ -1,9 +1,0 @@
-.footerWrapper {
-  width: 100%;
-  margin-top: 1rem;
-}
-
-.subTitle {
-  padding: 0 1rem;
-  font-weight: 400;
-}
